requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
azure-storage-blob
openai==0.28.1
numpy==1.24.4
pypdf2==3.0.1
azure-core==1.29.5
azure-search-documents==11.4.0
azure-ai-formrecognizer
beautifulsoup4
tiktoken
pycryptodome